+++ sentinal_ml/app/core/config.py
@@ -5,7 +5,7 @@ from dotenv import load_dotenv
 load_dotenv()
 
 class Settings:
-    PROJECT_NAME: str = os.getenv("PROJECT_NAME", "Sentinel AI")
+    PROJECT_NAME: str = os.getenv("PROJECT_NAME", "Sentinal AI")
 
     # Inside Docker, we will set this to /app/model/sentinel_model.joblib
     DEFAULT_MODEL_PATH = os.path.join(Path(__file__).resolve().parent.parent, "m_learning", "model", "sentinel_model.joblib")